Move platform-os/includes into sub directory /os
